feat: getters and types for standard `reg`, `ranges` and `dma-ranges` props (#19)
* Pass Fdt by value rather than reference.

It is Copy and only contains a single reference, so there's no point
passing it by reference. This also means that the lifetimes of things
returned from getters are only tied to the underlying slice, not the
Fdt itself, which may be more convenient.

* Add getter for standard `reg` property.

This requires keeping track of the #address-cells and #size-cells of the
parent node.

* Factor out Cells type for Reg.

* Make as_prop_encoded_array return array of Cells.

This saves work splitting the result in the caller.

* Add type and getters for standard ranges and dma-ranges properties.

* Fix test build for Rust 1.88.

* Fixes from code review.