refactor: validate `Fdt` on construction
This makes the APIs more ergonomic, as most of the methods that used to
return `Result<T, E>`, return `T` now instead, making it much easier to
work with, especially in case of iterators.

This obviously causes the constructor to run in linear time, which
shouldn't be a huge problem considering that any sensible use of the
parser will require linear-time traversal time anyway. In case this
would be a problem, however, a separate constructor that skips the
validation (and causes the traverse methods to panic instead) is added.

This also allows us to unify some error handling between in-memory and
internal representation APIs, bringing us closer to unifying these two APIs
under a common trait, and supporting the "standard nodes & properties"
for the internal representation variant.